Bug fixes on the Interface. UI Improvements.
Following the latest update, I’ve made improvements to the new design of the settings section. It still needs a lot of work, but I’m leaving it parked for now, as its visual state is acceptable. Back on the dashboard, I’ve added full support for mobile devices, both in web browsers and on iOS clients. I’m still refining some details, but so far, it’s looking very good.